Move LongExt to kotlinutil, rename to LongFormatting
